+++ site/index.tsx
@@ -1,5 +1,5 @@
 import { Dashboard } from "pages/Dashboard/Dashboard";
-import { Layout } from "pages/Layout";
+import { Layout } from "pages/layout";
 import { injectAnalytics } from "lib/injectAnalytics";
 import m from "mithril";
